+++ client/client.go
@@ -1,0 +1,151 @@
+package client
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"net/http"
+	"net/http/httputil"
+
+	cac "github.com/aptible/cloud-api-clients/clients/go"
+)
+
+// client - internal cac struct used only for this service with some common configuration
+type client struct {
+	ctx context.Context
+
+	apiClient *cac.APIClient
+	debug     bool
+	token     string
+}
+
+// NewClient - generate a new cloud api cloud_api_client
+func NewClient(debug bool, host string, token string) CloudClient {
+	config := cac.NewConfiguration()
+	config.Host = host
+	config.Scheme = "https"
+
+	apiClient := cac.NewAPIClient(config)
+
+	ctx := context.Background()
+	ctx = context.WithValue(ctx, cac.ContextAccessToken, token)
+
+	return &client{
+		ctx:       ctx,
+		apiClient: apiClient,
+
+		debug: debug,
+		token: token,
+	}
+}
+
+func (c *client) HandleResponse(r *http.Response) {
+	if r == nil {
+		fmt.Printf("The HTTP response is nil which means the request was never made.  Are you sure your API domain is set properly? (%s)\n", c.apiClient.GetConfig().Host)
+		return
+	}
+	c.PrintResponse(r)
+}
+
+func (c *client) PrintResponse(r *http.Response) {
+	if !c.debug {
+		return
+	}
+
+	log.Println("--- DEBUG ---")
+	reqDump, err := httputil.DumpRequestOut(r.Request, false)
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	log.Printf("REQUEST:\n%s", string(reqDump))
+
+	respDump, err := httputil.DumpResponse(r, true)
+	if err != nil {
+		log.Println(err)
+	}
+
+	log.Printf("RESPONSE:\n%s\n", string(respDump))
+}
+
+func (c *client) ListEnvironments(orgId string) ([]cac.EnvironmentOutput, error) {
+	request := c.apiClient.EnvironmentsApi.GetEnvironmentsApiV1OrganizationsOrganizationIdEnvironmentsGet(c.ctx, orgId)
+	env, r, err := request.Execute()
+	c.HandleResponse(r)
+	return env, err
+}
+
+func (c *client) CreateEnvironment(orgId string, params cac.EnvironmentInput) (*cac.EnvironmentOutput, error) {
+	request := c.apiClient.EnvironmentsApi.CreateEnvironmentApiV1OrganizationsOrganizationIdEnvironmentsPost(c.ctx, orgId).EnvironmentInput(params)
+	env, r, err := request.Execute()
+	c.HandleResponse(r)
+	return env, err
+}
+
+func (c *client) DestroyEnvironment(orgId string, envId string) error {
+	_, r, err := c.apiClient.EnvironmentsApi.DeleteEnvironmentByIdApiV1OrganizationsOrganizationIdEnvironmentsEnvironmentIdDelete(c.ctx, envId, orgId).Execute()
+	c.HandleResponse(r)
+	return err
+}
+
+func (c *client) CreateOrg(orgId string, params cac.OrganizationInput) (*cac.OrganizationOutput, error) {
+	request := c.apiClient.OrganizationsApi.PutOrganizationApiV1OrganizationsOrganizationIdPut(c.ctx, orgId).OrganizationInput(params)
+	org, r, err := request.Execute()
+	c.HandleResponse(r)
+	return org, err
+}
+
+func (c *client) FindOrg(orgId string) (*cac.OrganizationOutput, error) {
+	org, r, err := c.apiClient.OrganizationsApi.GetOrganizationByIdApiV1OrganizationsOrganizationIdGet(c.ctx, orgId).Execute()
+	c.HandleResponse(r)
+	return org, err
+}
+
+func (c *client) CreateAsset(orgId string, envId string, params cac.AssetInput) (*cac.AssetOutput, error) {
+	request := c.apiClient.AssetsApi.CreateAssetApiV1OrganizationsOrganizationIdEnvironmentsEnvironmentIdAssetsPost(c.ctx, envId, orgId).AssetInput(params)
+	asset, r, err := request.Execute()
+	c.HandleResponse(r)
+	return asset, err
+}
+
+func (c *client) DestroyAsset(orgId string, envId string, assetId string) error {
+	request := c.apiClient.AssetsApi.DeleteAssetByIdApiV1OrganizationsOrganizationIdEnvironmentsEnvironmentIdAssetsAssetIdDelete(c.ctx, assetId, envId, orgId)
+	_, r, err := request.Execute()
+	c.HandleResponse(r)
+	return err
+}
+
+func (c *client) ListAssets(orgId string, envId string) ([]cac.AssetOutput, error) {
+	request := c.apiClient.AssetsApi.GetAssetsApiV1OrganizationsOrganizationIdEnvironmentsEnvironmentIdAssetsGet(c.ctx, envId, orgId)
+	assets, r, err := request.Execute()
+	c.HandleResponse(r)
+	return assets, err
+}
+
+func (c *client) DescribeAsset(orgId string, envId string, assetId string) (*cac.AssetOutput, error) {
+	request := c.apiClient.AssetsApi.GetAssetByIdApiV1OrganizationsOrganizationIdEnvironmentsEnvironmentIdAssetsAssetIdGet(c.ctx, assetId, envId, orgId)
+	asset, r, err := request.Execute()
+	c.HandleResponse(r)
+	return asset, err
+}
+
+func (c *client) ListOrgs() ([]cac.OrganizationOutput, error) {
+	request := c.apiClient.OrganizationsApi.GetOrganizationsApiV1OrganizationsGet(c.ctx)
+	orgs, r, err := request.Execute()
+	c.HandleResponse(r)
+	return orgs, err
+}
+
+func (c *client) ListOperationsByAsset(orgId string, assetId string) ([]cac.OperationOutput, error) {
+	request := c.apiClient.OperationsApi.GetOperationsApiV1OrganizationsOrganizationIdOperationsGet(c.ctx, orgId).AssetId(assetId)
+	ops, r, err := request.Execute()
+	c.HandleResponse(r)
+	return ops, err
+}
+
+func (c *client) ListAssetBundles(orgId string, envId string) ([]cac.AssetBundle, error) {
+	request := c.apiClient.EnvironmentsApi.GetEnvironmentAllowedAssetsApiV1OrganizationsOrganizationIdEnvironmentsEnvironmentIdAssetBundlesGet(c.ctx, envId, orgId)
+	bundles, r, err := request.Execute()
+	c.HandleResponse(r)
+	return bundles, err
+}
