--- internal/common/constants.go
@@ -1,1 +1,0 @@
-package common
