+++ ui/common/view.go
